// src/app/business-logic/model/queues/queuesGetQueueMetricsRequest.ts
/**
 * queues
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.9
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface QueuesGetQueueMetricsRequest {
    /**
     * Starting time (in seconds from epoch) for collecting metrics
     */
    from_date: number;
    /**
     * Ending time (in seconds from epoch) for collecting metrics
     */
    to_date: number;
    /**
     * Time interval in seconds for a single metrics point. The minimal value is 1
     */
    interval: number;
    /**
     * List of queue ids to collect metrics for. If not provided or empty then all   then average metrics across all the company queues will be returned.
     */
    queue_ids?: Array<string>;
}