// src/app/business-logic/model/queues/queuesGetQueueMetricsResponse.ts
/**
 * queues
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.14
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { QueueMetrics } from '././queueMetrics';


export interface QueuesGetQueueMetricsResponse {
    /**
     * List of the requested queues with their metrics. If no queue ids were requested   then \'all\' queue is returned with the metrics averaged accross all the company   queues.
     */
    queues?: Array<QueueMetrics>;
}