// src/app/business-logic/model/queues/queueMetrics.ts
/**
 * queues
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.14
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface QueueMetrics {
    /**
     * ID of the queue
     */
    queue?: string;
    /**
     * List of timestamps (in seconds from epoch) in the acceding order. The timestamps are separated by the requested interval. Timestamps where no queue status change was recorded are omitted.
     */
    dates?: Array<number>;
    /**
     * List of average waiting times for tasks in the queue. The points correspond to the timestamps in the dates list. If more than one value exists for the given interval then the maximum value is taken.
     */
    avg_waiting_times?: Array<number>;
    /**
     * List of tasks counts in the queue. The points correspond to the timestamps in the dates list. If more than one value exists for the given interval then the count that corresponds to the maximum average value is taken.
     */
    queue_lengths?: Array<number>;
}